requirements.txt
vk_api
python-dotenv
pytest
